move functionality from pivot application to cube view
