+++ src/client/components/cube-view/cube-view.mocha.tsx
@@ -1,0 +1,30 @@
+'use strict';
+
+import { expect } from 'chai';
+import * as sinon from 'sinon';
+import { setupDOM } from '../../utils/jsdom-setup';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import * as TestUtils from 'react-addons-test-utils';
+
+import { $, Expression } from 'plywood';
+import { CubeView } from './cube-view';
+
+describe('CubeView', () => {
+  setupDOM();
+
+  it('adds the correct class', () => {
+    var renderedComponent = TestUtils.renderIntoDocument(
+      <CubeView essence={null}
+
+      />
+    );
+    console.log(renderedComponent);
+
+    expect(TestUtils.isCompositeComponent(renderedComponent), 'should be composite').to.equal(true);
+    expect((ReactDOM.findDOMNode(renderedComponent) as any).className, 'should contain class').to.contain('cube-view');
+
+  });
+});
+
+
