+++ src/client/components/cube-view/cube-view.scss
@@ -1,0 +1,5 @@
+@import '../../imports';
+
+.cube-view {
+
+}
